Added sensors and limits
Added an Adjuster functional interface, which takes a value and returns a value of the same type, adjusted.

Added DistanceSensor class, which extends AnalogInput and includes an Adjuster to convert voltage to distance, as well as methods to apply it and check if the distance is in a range. Also added a LaserSensor class extending it, and determinig its converter based on two points or slope and intercept.

Modified MotorSubsystem to have a Limit, to prevent it from mobving when it shouldn't, and added Limit interface and objects, to e implemented with the sysem.