fix(consts): add private addition with parameter, to prevent null keys
Since the constants given as value in addToDashboard are specifically initialized in the classes inheriting constants, when used in the constant constructor itself, it threw a NullPointerException.

To fix that, added a protected addToDashboard method which takes the value put as a parameter, to give the value directly to it construction instead of the not yet initialized inherited default.